+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Version History
 
+## 0.6.0 (Unreleased)
+
+### Library Changes
+
+  * A new module (`Edify.Compiler.Build`) has been added that provides
+    an easy-to-use `convert` function to convert Edify-flavored
+    Markdown to whatever Markdown variant you are using.
+
 ## 0.5.0 (March 2, 2021)
 
   * First documented version